requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
click==8.3.0
colorama==0.4.6
Django==5.2.8
django-ninja==1.4.5
django-schema-viewer
gunicorn==23.0.0
h11==0.16.0
httptools==0.7.1
idna==3.11
packaging==25.0
psycopg2-binary==2.9.11
pydantic==2.12.4
pydantic[email]
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
sniffio==1.3.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
ortools==9.10.4067
matplotlib==3.9.0
python-dateutil